core: LogDrainLoop in its own header
The drain thread is self-contained (owned by Core::State, started by
start_log_drain_locked, stopped by whoever takes it out of the state) and Core.h already
treats it as its own thing with a forward declaration, so it lives in
src/scheduler/LogDrainLoop.h and Core.cpp opens with the state and the lifecycle.
