+++ CHANGELOG.md
@@ -21,7 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **Breaking (2.x classes):** the process-wide singleton the 2.x handler registers with is the *core*: `anira::Context` is `anira::Core` (`anira/scheduler/Core.h`; its private state struct is `Core::State`), `anira::ContextConfig` is `anira::CoreConfig` (`anira/CoreConfig.h`, with `LogConfig` / `LogDrain` / `WaitStrategy` / `LogLevel` unchanged), `Context::register_machine` is `Core::register_context`, `InferenceManager::get_context()` is `get_core()`, `JsonConfigLoader::get_context_config()` is `get_core_config()`, the bridge's `to_context_config` is `to_core_config`, and the log group `anira.context` is `anira.core`. The 2.x JSON key `context_config` is unchanged (it is the document format). Migration: the handler used to take a `ContextConfig`; in the 3.x runtime it takes an `anira_context*` built from an `anira_context_config`, and until then it takes a `CoreConfig`, which is the same struct under its truthful name. `Core::shutdown()` and `Core::make_inference_thread()` are the two members the guides showed; `anira_shutdown` and `anira_inference_thread_create` are their 3.x forms.
 - **Breaking:** `anira_shutdown` (and the CLAP example's `clap_deinit`, which now calls it) is effective only when no context and no handler exist in this copy of anira; otherwise nothing happens and it returns `ANIRA_ERROR_INVALID_STATE`, so one client of a shared `libanira` cannot silence another's sessions. `anira::Core::shutdown()`'s unconditional form stays internal for the library-unload hook. Migration: destroy every context and handler before the module-exit entry point runs, which a plugin does anyway.
 - **Breaking:** `anira::CoreConfig::m_anira_version` and `m_enabled_backends` are gone, with the cross-session version and enabled-backends compares of the context (inside one copy every session shares one header; across two copies nothing can see the other). What the build compiled in is `anira_enabled_backends`; what is usable here is `anira_capabilities_backends` on a probed context. Migration: nothing to write; code that read the two fields uses the two entries.
-- The core owns the real-time log drain thread: a low-priority `thl::core::Thread` named `anira-log` of anira's own runs `Queue::drain()` every drain interval while a session or a context exists (tanh-lib's `DrainThread` is no longer constructed), and the last user's release, `anira_context_destroy` or `anira_shutdown` stops it and flushes the queue through the sinks on the calling thread. `ANIRA_LOG_FLAG_TRACE_FAILURES` is admitted by `anira_context_config_set_log_flags` and the `flags` of `anira_log_desc`.
+- The core owns the real-time log drain thread: a low-priority `thl::core::Thread` named `anira-log` of anira's own (`LogDrainLoop`, `src/scheduler/LogDrainLoop.h`) runs `Queue::drain()` every drain interval while a session or a context exists (tanh-lib's `DrainThread` is no longer constructed), and the last user's release, `anira_context_destroy` or `anira_shutdown` stops it and flushes the queue through the sinks on the calling thread. `ANIRA_LOG_FLAG_TRACE_FAILURES` is admitted by `anira_context_config_set_log_flags` and the `flags` of `anira_log_desc`.
 - Every input and output ring of the pipeline is an instantiation of `RingBufferT<T>` with `T` the element type of the slot's ring dtype: `anira_ring` (`anira/utils/RingBuffer.h`, the type `anira::RingBuffer` now names) holds one instantiation per scalar dtype the ABI pins (float32, float64, float16, bfloat16, int8, uint8, bool8, int16, int32, int64, with `T` the dtype's C type and the three that have none stored as their bits; no two dtypes share a ring; float32 unless a slot says otherwise), chosen at prepare: `SessionElement::prepare` and `Core::prepare_session` take the per-slot `RingDtypes` and the caller's `CustomLatencies` (the former `std::vector<long>`, named), which the 2.x `InferenceManager::prepare` fills from its arguments (a new overload carries both for the 3.x prepare). The ring's block API (`push_block`, `pop_block`, `peek_past_block`, `push_fill`, `push_zeros`, `discard`, `pop_windows`, which now holds the batched sliding-window pop of `PrePostProcessor::pop_samples_from_buffer`) takes the caller's dtype, returns 0 and writes nothing on a disagreement, and never converts; the 2.x float API stays as the float32 face, so every 2.x pre/post processor compiles unchanged. Storage only: the typed Hard entries that reach the rings follow with the 3.x handler.
 
 ## [v3.0.0-alpha.1] - 2026-09-03
